Extract yubikey_conditions module
