+++ test/on_yubikey/cli_piv/test_fips.py
@@ -1,6 +1,7 @@
 from ykman.util import TRANSPORT
+from .. import yubikey_conditions
 from ...util import open_file
-from ..util import cli_test_suite, is_fips, DestructiveYubikeyTestCase
+from ..util import cli_test_suite, DestructiveYubikeyTestCase
 
 
 @cli_test_suite(TRANSPORT.CCID)
@@ -15,12 +16,12 @@ def additional_tests(ykman_cli):
         def tearDownClass(cls):
             ykman_cli('piv', 'reset', '-f')
 
-        @is_fips
+        @yubikey_conditions.is_fips
         def test_rsa1024_generate_blocked(self):
             with self.assertRaises(SystemExit):
                 ykman_cli('piv', 'generate-key', '9a', '-a', 'RSA1024', '-')
 
-        @is_fips
+        @yubikey_conditions.is_fips
         def test_rsa1024_import_blocked(self):
             with self.assertRaises(SystemExit):
                 with open_file('rsa_1024_key.pem') as f:
