+++ test/on_yubikey/test_cli_misc.py
@@ -1,7 +1,8 @@
 import unittest
 
 from ykman.util import TRANSPORT
-from .util import (cli_test_suite, is_fips, is_not_fips)
+from . import yubikey_conditions
+from .util import (cli_test_suite)
 
 
 @cli_test_suite(sum(TRANSPORT))
@@ -14,12 +15,12 @@ def additional_tests(ykman_cli):
             self.assertIn('Serial number:', info)
             self.assertIn('Firmware version:', info)
 
-        @is_not_fips
+        @yubikey_conditions.is_not_fips
         def test_ykman_info_does_not_report_fips_for_non_fips_device(self):
             info = ykman_cli('info', '--check-fips')
             self.assertNotIn('FIPS', info)
 
-        @is_fips
+        @yubikey_conditions.is_fips
         def test_ykman_info_reports_fips_status(self):
             info = ykman_cli('info', '--check-fips')
             self.assertIn('FIPS Approved Mode:', info)
